Refactor: rearranged module file structure
